% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaplaceMechanism_DP.R
\name{differential_privacy}
\alias{differential_privacy}
\title{Differential Privacy Implementation}
\usage{
differential_privacy(res, epsilon)
}
\arguments{
\item{res}{A list containing the original value (`og_value`) and 
the L1 sensitivity (`l1_sens`). For numeric vectors, `l1_sens` should be a single value. 
For matrices, `l1_sens` should be a vector with L1 sensitivities for each cell 
of the matrix. The `og_value` can be either a numeric vector or a matrix.}

\item{epsilon}{The privacy budget parameter. A non-zero value 
specifies the degree of noise to be added for privacy. 
Setting the value to NULL means no noise is added.}
}
\value{
The function returns the data with Laplace noise added for differential privacy. 
The structure of the returned data (matrix or numeric vector) matches that of the 
input `og_value`.
}
\description{
Applies differential privacy to a given dataset by adding Laplace noise.
The function supports both numeric vectors and matrices.
}
\details{
This function implements differential privacy by adding Laplace noise to 
the original data (`og_value`). The amount of noise depends on the L1 
sensitivity (`l1_sens`) of the query and the privacy budget (`epsilon`).
For numeric vectors, a single noise value is generated and added. 
For matrices, individual noise values are generated for each cell based on 
the provided `l1_sens` vector and added to the corresponding cells. 
The function uses `Laplace_noise_generator` to generate the required noise.
}
\examples{
# Example usage for a numeric vector
res <- list(og_value = c(1, 2, 3), l1_sens = 1)
differential_privacy(res, 0.5)

# Example usage for a matrix
matrix_sensitivities <- rep(2, 4) # Replace with actual sensitivities
res <- list(og_value = matrix(c(1, 2, 3, 4), nrow = 2), l1_sens = matrix_sensitivities)
differential_privacy(res, 0.5)
}
