% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.calcMCR.R
\name{ds.calcMCR}
\alias{ds.calcMCR}
\title{Calculate missing classification rate (MCR)}
\usage{
ds.calcMCR(ws = ws, datasourceTest, X, Y, average = T)
}
\arguments{
\item{ws}{The list of coefficient vectors for all servers}

\item{datasourceTest}{The connection of the target server}

\item{X}{The name of design matrices}

\item{Y}{The name of labels}

\item{average}{The indicator to average or not the prediction probability over all coefficient vectors}
}
\value{
The MCR of each coefficient vectors (or averaged MCR)
}
\description{
Calculate missing classification rate (MCR) on the target server
}
\details{
Calculate missing classification rate (MCR) on the target server
}
\author{
Han Cao
}
