% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.calcMSE.R
\name{ds.calcMSE}
\alias{ds.calcMSE}
\title{Calculate mean-squared error (MSE)}
\usage{
ds.calcMSE(ws, datasourceTest, X, Y, average = T)
}
\arguments{
\item{ws}{The list of coefficient vectors for all servers}

\item{datasourceTest}{The connection of the target server}

\item{X}{The name of design matrices}

\item{Y}{The name of labels}

\item{average}{The indicator to average or not the prediction probability over all coefficient vectors}
}
\value{
The MSE of each coefficient vector (or averaged MSE)
}
\description{
Calculate mean-squared error (MSE) on the target server
}
\details{
Calculate mean-squared error (MSE) on the target server
}
\author{
Han Cao
}
