% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.LassoCov.R
\name{ds.LassoCov_CVInSite}
\alias{ds.LassoCov_CVInSite}
\title{In-site cross-validation with the control of covariates}
\usage{
ds.LassoCov_CVInSite(
  X = NULL,
  Y = NULL,
  type = "regress",
  nfolds = 10,
  lam_ratio = 0.01,
  nlambda = 10,
  lambda = NULL,
  opts = list(init = 0, maxIter = 50, tol = 0.01, ter = 2),
  covar = NULL,
  datasources = NULL,
  nDigits = 10,
  intercept = F
)
}
\arguments{
\item{X}{The design matrices of multiple cohorts}

\item{Y}{Label vectors of multiple cohorts (in case of a binary outcome classes must to be coded as -1 and 1)}

\item{type}{regression(=regress) or classification(=classify)}

\item{nfolds}{The number of folds}

\item{lam_ratio}{smallest lambda / largest lambda}

\item{nlambda}{The length of lambda sequence}

\item{lambda}{The lambda sequence}

\item{opts}{Options controlling the optimization procedure}

\item{covar}{Positions of adjusting covariates in the X dataset}

\item{datasources}{The connections of servers}

\item{nDigits}{The number of digits rounded for each number prepared for network transmission}

\item{intercept}{Use intercept(=TRUE) or non-intercept(=FALSE) model}
}
\value{
The result of cross-validation
}
\description{
In-site cross-validation
}
\details{
Cross-site cross-validation
}
\author{
Han Cao & Augusto Anguita-Ruiz
}
