% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.lmBetas.R
\name{ds.lmBetas}
\alias{ds.lmBetas}
\title{Linear regression model to selected adjusting covariates}
\usage{
ds.lmBetas(X, Y, covar)
}
\arguments{
\item{X}{Predictors}

\item{Y}{Outcome}

\item{covar}{Position of adjusting covariates in the predictor dataset}
}
\value{
Estimated beta coefficients
}
\description{
Apply linear regression model to selected adjusting covariates
}
\author{
Augusto Anguita-Ruiz & Han Cao
}
