% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.Lasso.R
\name{ds.LR_Lasso}
\alias{ds.LR_Lasso}
\title{Solver of logistic regression with Lasso}
\usage{
ds.LR_Lasso(X, Y, lam, C, opts, datasources, nDigits)
}
\arguments{
\item{X}{The design matrices of multiple cohorts}

\item{Y}{Label vectors of multiple cohorts}

\item{lam}{The hyper-parameter controlling the sparsity}

\item{C}{The hyper-parameter associated with L2 term}

\item{opts}{Options controlling the optimization procedure}

\item{datasources}{The connections of servers}

\item{nDigits}{The number of digits rounded for each number prepared for network transmission}
}
\value{
The converged result of optimization
}
\description{
Solver of logistic regression with Lasso
}
\details{
Solver of logistic regression with Lasso
}
\author{
Han Cao
}
