% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.MTL_L21.R
\name{ds.LS_MTL_L21}
\alias{ds.LS_MTL_L21}
\title{Solver of FeMTL with least-square loss for joint feature selection}
\usage{
ds.LS_MTL_L21(X, Y, lam, C, opts, datasources = NULL, nDigits)
}
\arguments{
\item{X}{The design matrices of multiple cohorts}

\item{Y}{Label vectors of multiple cohorts}

\item{lam}{The hyper-parameter controlling the sparsity}

\item{C}{The hyper-parameter associated with L2 term}

\item{opts}{Options controlling the optimization procedure}

\item{datasources}{The connections of servers}

\item{nDigits}{The number of digits rounded for each number prepared for network transmission}
}
\value{
The converged result of optimization
}
\description{
Solver of FeMTL with least-square loss for joint feature selection
}
\details{
Solver of FeMTL with least-square loss for joint feature selection
}
\author{
Han Cao
}
