% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.MTL_Net.R
\name{ds.LS_MTL_Net}
\alias{ds.LS_MTL_Net}
\title{Solver of FeMTL with least-square loss and network incorporation}
\usage{
ds.LS_MTL_Net(X, Y, lam, C, G, opts, datasources = NULL, nDigits)
}
\arguments{
\item{X}{The design matrices of multiple cohorts}

\item{Y}{Label vectors of multiple cohorts}

\item{lam}{The hyper-parameter controlling the sparsity}

\item{C}{The hyper-parameter associated with network constraint}

\item{G}{The graph described the task relatedness}

\item{opts}{Options controlling the optimization procedure}

\item{datasources}{The connections of servers}

\item{nDigits}{The number of digits rounded for each number prepared for network transmission}
}
\value{
The converged result of optimization
}
\description{
Solver of FeMTL with least-square loss and network incorporation
}
\details{
Solver of FeMTL with least-square loss and network incorporation
}
\author{
Han Cao
}
