% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.MTL_iNMF.R
\name{ds.MTL_iNMF_Train}
\alias{ds.MTL_iNMF_Train}
\title{Training models of dsMTL_iNMF}
\usage{
ds.MTL_iNMF_Train(
  datasources,
  Xs,
  rank,
  myServerKey = NULL,
  n_initializations,
  Sp,
  lam,
  opts,
  nDigits = 10
)
}
\arguments{
\item{datasources}{The connections of servers}

\item{Xs}{The data matrices of multiple cohorts}

\item{rank}{The hyper-parameter indicate the number of  latent variables}

\item{myServerKey}{The key was granted by the server administrator to access the raw data}

\item{n_initializations}{The number of variables' initializations}

\item{Sp}{The hyper-parameter controlling the sparsity}

\item{lam}{The hyper-parameter associated with L2 term}

\item{opts}{Options controlling the optimization procedure}

\item{nDigits}{The number of digits rounded for each number prepared for network transmission}
}
\value{
The converged result of optimization
}
\description{
Training models of dsMTL_iNMF
}
\details{
Training models of dsMTL_iNMF
}
\author{
Han Cao
}
