% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.MTL_L21.R
\name{ds.MTL_L21_CVInSite}
\alias{ds.MTL_L21_CVInSite}
\title{In-site cross-validation for dsMTL_L21}
\usage{
ds.MTL_L21_CVInSite(
  X = NULL,
  Y = NULL,
  type = "regress",
  nfolds = 10,
  lam_ratio = 0.01,
  nlambda = 10,
  lambda = NULL,
  opts = list(init = 0, maxIter = 50, tol = 0.01, ter = 2),
  C = 0,
  datasources = NULL,
  nDigits = 10,
  intercept = F
)
}
\arguments{
\item{X}{The design matrices of multiple cohorts}

\item{Y}{Label vectors of multiple cohorts}

\item{type}{regression(=regress) or classification(=classify)}

\item{nfolds}{The number of folds}

\item{lam_ratio}{smallest lambda / largest lambda}

\item{nlambda}{The length of lambda sequence}

\item{lambda}{The lambda sequence}

\item{opts}{Options controlling the optimization procedure}

\item{C}{The hyper-parameter associated with L2 term}

\item{datasources}{The connections of servers}

\item{nDigits}{The number of digits rounded for each number prepared for network transmission}

\item{intercept}{Use intercept(=TRUE) or non-intercept(=FALSE) model}
}
\value{
The result of cross-validation
}
\description{
In-site cross-validation for dsMTL_L21
}
\details{
In-site cross-validation for dsMTL_L21
}
\author{
Han Cao
}
