% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.setMyServerData.R
\name{ds.setMyServerData}
\alias{ds.setMyServerData}
\title{Set data matrix on server}
\usage{
ds.setMyServerData(myServerKey, data, symbol)
}
\arguments{
\item{myServerKey}{The identify of the access. myServerKey$server contained a valid server connection. myServerKey$key contained a valid key.}

\item{data}{The name of the target matrix on the server}

\item{symbol}{The name of the new variable}
}
\value{
The requested matrix or error message
}
\description{
set data matrix from the valid client
}
\details{
The "Datakey" mechanism allowed the valid client upload the data matrix directly into the memory of the server. The validity 
was granted by the server administrator. To achieve this, the administrator generated a key and put the local copy in the dataset 
serverDataKey.myKey(project: serverDataKey; table: myKey) where the dataset was created normaly using DataSHIELD. The remote ropy
of the key was hand over to the trustworthy analyst. This was determined by the server administrator
}
\author{
Han Cao
}
