% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.subsetSubjests.R
\name{ds.subsetSubjests}
\alias{ds.subsetSubjests}
\title{Subset subjects for cross-validation}
\usage{
ds.subsetSubjests(datasources, idx, newSymbol, symbol)
}
\arguments{
\item{datasources}{The connections of servers}

\item{idx}{the index of subjects}

\item{newSymbol}{the name of new variable}

\item{symbol}{the name of raw data}
}
\description{
Subset subjects for cross-validation
}
\details{
In each server, the subjects of index were selected and copied into a new symbol
}
\author{
Han Cao
}
