% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.lmBetas.R
\name{ds.lmBetas}
\alias{ds.lmBetas}
\title{Fit a linear regression model on selected covariates}
\usage{
ds.lmBetas(X, Y, covar, datasources)
}
\arguments{
\item{X}{Predictors}

\item{Y}{Outcome}

\item{covar}{Positions of adjusting covariates in the X dataset}

\item{datasources}{The connections of servers}
}
\value{
Estimated beta coefficients for covariates
}
\description{
Fit a linear regression model with only selected adjusting covariates
}
\details{
Beta coefficients are employed for the estimation of lambda max
}
\author{
Han Cao & Augusto Anguita-Ruiz
}
