% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.LassoCov.R
\name{ds.LS_LassoCov}
\alias{ds.LS_LassoCov}
\title{Solver of LassoCov}
\usage{
ds.LS_LassoCov(X, Y, lam, covar = NULL, opts, datasources, nDigits)
}
\arguments{
\item{X}{The design matrices of multiple cohorts}

\item{Y}{Label vectors of multiple cohorts (in case of a binary outcome classes must to be coded as -1 and 1)}

\item{lam}{The hyper-parameter controlling the sparsity}

\item{covar}{Positions of adjusting covariates in the X dataset}

\item{opts}{Options controlling the optimization procedure}

\item{datasources}{The connections of servers}

\item{nDigits}{The number of digits rounded for each number prepared for network transmission}
}
\value{
The converged result of optimization
}
\description{
Solver of Lasso regression with controlling covariate effect
}
\details{
Solver of Lasso regression with controlling covariate effect
}
\author{
Han Cao & Augusto Anguita-Ruiz
}
